Unify strtobool implementation
