+++ tools/alerts/create_alerts.py
@@ -12,6 +12,8 @@ from typing import Any
 
 import requests
 
+from tools.strtobool import strtobool
+
 
 ALL_SKIPPED_THRESHOLD = 100
 SIMILARITY_THRESHOLD = 0.75
@@ -63,22 +65,6 @@ DISABLED_ALERTS = [
     "rerun_disabled_tests",
     "unstable",
 ]
-
-
-def strtobool(val: str) -> bool:
-    """Convert a string representation of truth to true (1) or false (0).
-
-    True values are "y", "yes", "t", "true", "on", and "1"; false values
-    are "n", "no", "f", "false", "off", and "0".  Raises ValueError if
-    "val" is anything else.
-    """
-    val = val.lower()
-    if val in ("y", "yes", "t", "true", "on", "1"):
-        return True
-    elif val in ("n", "no", "f", "false", "off", "0"):
-        return False
-    else:
-        raise ValueError(f"invalid truth value {val!r}")
 
 
 class JobStatus:
